fastapi-auth-service/requirements.txt
fastapi
uvicorn
python-dotenv
bcrypt
sqlalchemy
pydantic
python-jose[cryptography]
authlib